Normalize Elasticsearch transport common javaagent package
