+++ instrumentation/elasticsearch/elasticsearch-transport-5.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/elasticsearch/transport/v5_0/Elasticsearch5TransportClientTest.java
@@ -8,7 +8,7 @@ package io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.v5_0;
 import static java.util.Collections.singletonList;
 import static org.elasticsearch.cluster.ClusterName.CLUSTER_NAME_SETTING;
 
-import io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.AbstractElasticsearchTransportClientTest;
+import io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.common.AbstractElasticsearchTransportClientTest;
 import java.io.File;
 import java.util.UUID;
 import org.elasticsearch.client.transport.TransportClient;
