+++ instrumentation/elasticsearch/elasticsearch-transport-6.0/testing/src/main/java/io/opentelemetry/javaagent/instrumentation/elasticsearch/transport/v6_0/AbstractElasticsearch6NodeClientTest.java
@@ -8,7 +8,7 @@ package io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.v6_0;
 import static java.util.Collections.singletonMap;
 import static org.elasticsearch.cluster.ClusterName.CLUSTER_NAME_SETTING;
 
-import io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.AbstractElasticsearchNodeClientTest;
+import io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.common.AbstractElasticsearchNodeClientTest;
 import java.io.File;
 import java.util.UUID;
 import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
