+++ instrumentation/elasticsearch/elasticsearch-transport-common/testing/src/main/java/io/opentelemetry/javaagent/instrumentation/elasticsearch/transport/common/AbstractElasticsearchTransportClientTest.java
@@ -3,17 +3,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.elasticsearch.transport;
+package io.opentelemetry.javaagent.instrumentation.elasticsearch.transport.common;
 
 import static io.opentelemetry.api.common.AttributeKey.longKey;
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 import static io.opentelemetry.instrumentation.api.internal.SemconvStability.emitStableDatabaseSemconv;
-import static io.opentelemetry.instrumentation.testing.junit.db.DbClientMetricsTestUtil.assertDurationMetric;
 import static io.opentelemetry.instrumentation.testing.junit.db.SemconvStabilityUtil.maybeStable;
+import static io.opentelemetry.instrumentation.testing.util.TelemetryDataUtil.orderByRootSpanName;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
-import static io.opentelemetry.semconv.DbAttributes.DB_OPERATION_NAME;
-import static io.opentelemetry.semconv.DbAttributes.DB_SYSTEM_NAME;
+import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.satisfies;
 import static io.opentelemetry.semconv.ErrorAttributes.ERROR_TYPE;
+import static io.opentelemetry.semconv.ExceptionAttributes.EXCEPTION_MESSAGE;
+import static io.opentelemetry.semconv.ExceptionAttributes.EXCEPTION_STACKTRACE;
+import static io.opentelemetry.semconv.ExceptionAttributes.EXCEPTION_TYPE;
+import static io.opentelemetry.semconv.NetworkAttributes.NETWORK_PEER_ADDRESS;
+import static io.opentelemetry.semconv.NetworkAttributes.NETWORK_PEER_PORT;
+import static io.opentelemetry.semconv.NetworkAttributes.NETWORK_TYPE;
 import static io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DB_OPERATION;
 import static io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DB_SYSTEM;
 import static io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DbSystemNameIncubatingValues.ELASTICSEARCH;
@@ -36,6 +41,7 @@ import org.elasticsearch.action.index.IndexResponse;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.cluster.health.ClusterHealthStatus;
 import org.elasticsearch.index.IndexNotFoundException;
+import org.elasticsearch.transport.RemoteTransportException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -44,7 +50,12 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 @SuppressWarnings("deprecation") // using deprecated semconv
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasticsearchClientTest {
+public abstract class AbstractElasticsearchTransportClientTest
+    extends AbstractElasticsearchClientTest {
+
+  protected abstract String getAddress();
+
+  protected abstract int getPort();
 
   private Stream<Arguments> healthArguments() {
     return Stream.of(
@@ -75,18 +86,29 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
                         .hasKind(SpanKind.CLIENT)
                         .hasParent(trace.getSpan(0))
                         .hasAttributesSatisfyingExactly(
-                            equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
-                            equalTo(maybeStable(DB_OPERATION), "ClusterHealthAction"),
-                            equalTo(
-                                stringKey("elasticsearch.action"),
-                                experimental("ClusterHealthAction")),
-                            equalTo(
-                                stringKey("elasticsearch.request"),
-                                experimental("ClusterHealthRequest"))),
+                            addNetworkTypeAttribute(
+                                equalTo(NETWORK_PEER_ADDRESS, getAddress()),
+                                equalTo(NETWORK_PEER_PORT, getPort()),
+                                equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
+                                equalTo(maybeStable(DB_OPERATION), "ClusterHealthAction"),
+                                equalTo(
+                                    stringKey("elasticsearch.action"),
+                                    experimental("ClusterHealthAction")),
+                                equalTo(
+                                    stringKey("elasticsearch.request"),
+                                    experimental("ClusterHealthRequest")))),
                 span ->
                     span.hasName("callback")
                         .hasKind(SpanKind.INTERNAL)
                         .hasParent(trace.getSpan(0))));
+  }
+
+  private List<AttributeAssertion> addNetworkTypeAttribute(AttributeAssertion... assertions) {
+    List<AttributeAssertion> result = new ArrayList<>(asList(assertions));
+    if (hasNetworkType()) {
+      result.add(satisfies(NETWORK_TYPE, val -> val.isIn("ipv4", "ipv6")));
+    }
+    return result;
   }
 
   private Stream<Arguments> errorArguments() {
@@ -121,7 +143,7 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
                     stringKey("elasticsearch.request.indices"), experimental("invalid-index"))));
 
     if (emitStableDatabaseSemconv()) {
-      assertions.add(equalTo(ERROR_TYPE, "org.elasticsearch.index.IndexNotFoundException"));
+      assertions.add(equalTo(ERROR_TYPE, "org.elasticsearch.transport.RemoteTransportException"));
     }
 
     testing.waitAndAssertTraces(
@@ -138,7 +160,20 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
                         .hasKind(SpanKind.CLIENT)
                         .hasParent(trace.getSpan(0))
                         .hasStatus(StatusData.error())
-                        .hasException(expectedException)
+                        .hasEventsSatisfyingExactly(
+                            event ->
+                                event
+                                    .hasName("exception")
+                                    .hasAttributesSatisfyingExactly(
+                                        equalTo(
+                                            EXCEPTION_TYPE,
+                                            RemoteTransportException.class.getName()),
+                                        satisfies(
+                                            EXCEPTION_MESSAGE,
+                                            val -> val.contains("indices:data/read/get")),
+                                        satisfies(
+                                            EXCEPTION_STACKTRACE,
+                                            val -> val.contains("IndexNotFoundException"))))
                         .hasAttributesSatisfyingExactly(assertions),
                 span ->
                     span.hasName("callback")
@@ -146,28 +181,8 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
                         .hasParent(trace.getSpan(0))));
   }
 
-  protected void waitYellowStatus() {
-    client()
-        .admin()
-        .cluster()
-        .prepareHealth()
-        .setWaitForYellowStatus()
-        .execute()
-        .actionGet(TIMEOUT);
-  }
-
-  private static List<AttributeAssertion> getActionAttributes(
-      String indexName, String indexType, String id, long version) {
-    return new ArrayList<>(
-        asList(
-            equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
-            equalTo(maybeStable(DB_OPERATION), "GetAction"),
-            equalTo(stringKey("elasticsearch.action"), experimental("GetAction")),
-            equalTo(stringKey("elasticsearch.request"), experimental("GetRequest")),
-            equalTo(stringKey("elasticsearch.request.indices"), experimental(indexName)),
-            equalTo(stringKey("elasticsearch.type"), experimental(indexType)),
-            equalTo(stringKey("elasticsearch.id"), experimental(id)),
-            equalTo(longKey("elasticsearch.version"), experimental(version))));
+  protected String getPutMappingActionName() {
+    return "PutMappingAction";
   }
 
   @Test
@@ -180,7 +195,6 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
     CreateIndexResponse indexResult = client.admin().indices().prepareCreate(indexName).get();
     assertThat(indexResult.isAcknowledged()).isTrue();
 
-    waitYellowStatus();
     GetResponse emptyResult = client.prepareGet(indexName, indexType, id).get();
     assertThat(emptyResult.isExists()).isFalse();
     assertThat(emptyResult.getId()).isEqualTo(id);
@@ -201,7 +215,10 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
     assertThat(result.getType()).isEqualTo(indexType);
     assertThat(result.getIndex()).isEqualTo(indexName);
 
-    testing.waitAndAssertTraces(
+    // PutMappingAction and IndexAction run in separate threads so their order can vary
+    testing.waitAndAssertSortedTraces(
+        orderByRootSpanName(
+            "CreateIndexAction", getPutMappingActionName(), "IndexAction", "GetAction"),
         trace ->
             trace.hasSpansSatisfyingExactly(
                 span ->
@@ -209,32 +226,73 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
                         .hasKind(SpanKind.CLIENT)
                         .hasNoParent()
                         .hasAttributesSatisfyingExactly(
-                            equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
-                            equalTo(maybeStable(DB_OPERATION), "CreateIndexAction"),
-                            equalTo(
-                                stringKey("elasticsearch.action"),
-                                experimental("CreateIndexAction")),
-                            equalTo(
-                                stringKey("elasticsearch.request"),
-                                experimental("CreateIndexRequest")),
-                            equalTo(
-                                stringKey("elasticsearch.request.indices"),
-                                experimental(indexName)))),
+                            addNetworkTypeAttribute(
+                                equalTo(NETWORK_PEER_ADDRESS, getAddress()),
+                                equalTo(NETWORK_PEER_PORT, getPort()),
+                                equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
+                                equalTo(maybeStable(DB_OPERATION), "CreateIndexAction"),
+                                equalTo(
+                                    stringKey("elasticsearch.action"),
+                                    experimental("CreateIndexAction")),
+                                equalTo(
+                                    stringKey("elasticsearch.request"),
+                                    experimental("CreateIndexRequest")),
+                                equalTo(
+                                    stringKey("elasticsearch.request.indices"),
+                                    experimental(indexName))))),
         trace ->
             trace.hasSpansSatisfyingExactly(
                 span ->
-                    span.hasName("ClusterHealthAction")
+                    span.hasName(getPutMappingActionName())
                         .hasKind(SpanKind.CLIENT)
                         .hasNoParent()
                         .hasAttributesSatisfyingExactly(
                             equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
-                            equalTo(maybeStable(DB_OPERATION), "ClusterHealthAction"),
+                            equalTo(maybeStable(DB_OPERATION), getPutMappingActionName()),
                             equalTo(
                                 stringKey("elasticsearch.action"),
-                                experimental("ClusterHealthAction")),
+                                experimental(getPutMappingActionName())),
                             equalTo(
                                 stringKey("elasticsearch.request"),
-                                experimental("ClusterHealthRequest")))),
+                                experimental("PutMappingRequest")))),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span ->
+                    span.hasName("IndexAction")
+                        .hasKind(SpanKind.CLIENT)
+                        .hasNoParent()
+                        .hasAttributesSatisfyingExactly(
+                            addNetworkTypeAttribute(
+                                equalTo(NETWORK_PEER_ADDRESS, getAddress()),
+                                equalTo(NETWORK_PEER_PORT, getPort()),
+                                equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
+                                equalTo(maybeStable(DB_OPERATION), "IndexAction"),
+                                equalTo(
+                                    stringKey("elasticsearch.action"), experimental("IndexAction")),
+                                equalTo(
+                                    stringKey("elasticsearch.request"),
+                                    experimental("IndexRequest")),
+                                equalTo(
+                                    stringKey("elasticsearch.request.indices"),
+                                    experimental(indexName)),
+                                equalTo(
+                                    stringKey("elasticsearch.request.write.type"),
+                                    experimental(indexType)),
+                                equalTo(
+                                    longKey("elasticsearch.response.status"), experimental(201)),
+                                equalTo(
+                                    longKey("elasticsearch.shard.replication.total"),
+                                    experimental(2)),
+                                equalTo(
+                                    longKey("elasticsearch.shard.replication.successful"),
+                                    experimental(1)),
+                                equalTo(
+                                    longKey("elasticsearch.shard.replication.failed"),
+                                    experimental(0)),
+                                equalTo(
+                                    longKey("elasticsearch.request.write.version"),
+                                    hasWriteVersion() ? experimental(-3) : null)))),
+        // moved here by sorting, chronologically happens before PutMappingAction
         trace ->
             trace.hasSpansSatisfyingExactly(
                 span ->
@@ -246,35 +304,6 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
         trace ->
             trace.hasSpansSatisfyingExactly(
                 span ->
-                    span.hasName("IndexAction")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasNoParent()
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
-                            equalTo(maybeStable(DB_OPERATION), "IndexAction"),
-                            equalTo(stringKey("elasticsearch.action"), experimental("IndexAction")),
-                            equalTo(
-                                stringKey("elasticsearch.request"), experimental("IndexRequest")),
-                            equalTo(
-                                stringKey("elasticsearch.request.indices"),
-                                experimental(indexName)),
-                            equalTo(
-                                stringKey("elasticsearch.request.write.type"),
-                                experimental(indexType)),
-                            equalTo(longKey("elasticsearch.response.status"), experimental(201)),
-                            equalTo(
-                                longKey("elasticsearch.shard.replication.total"), experimental(2)),
-                            equalTo(
-                                longKey("elasticsearch.shard.replication.successful"),
-                                experimental(1)),
-                            equalTo(
-                                longKey("elasticsearch.shard.replication.failed"), experimental(0)),
-                            equalTo(
-                                longKey("elasticsearch.request.write.version"),
-                                hasWriteVersion() ? experimental(-3) : null))),
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
                     span.hasName("GetAction")
                         .hasKind(SpanKind.CLIENT)
                         .hasNoParent()
@@ -282,17 +311,27 @@ public abstract class AbstractElasticsearchNodeClientTest extends AbstractElasti
                             getActionAttributes(indexName, indexType, id, 1))));
   }
 
+  private List<AttributeAssertion> getActionAttributes(
+      String indexName, String indexType, String id, long version) {
+    return new ArrayList<>(
+        addNetworkTypeAttribute(
+            equalTo(NETWORK_PEER_ADDRESS, getAddress()),
+            equalTo(NETWORK_PEER_PORT, getPort()),
+            equalTo(maybeStable(DB_SYSTEM), ELASTICSEARCH),
+            equalTo(maybeStable(DB_OPERATION), "GetAction"),
+            equalTo(stringKey("elasticsearch.action"), experimental("GetAction")),
+            equalTo(stringKey("elasticsearch.request"), experimental("GetRequest")),
+            equalTo(stringKey("elasticsearch.request.indices"), experimental(indexName)),
+            equalTo(stringKey("elasticsearch.type"), experimental(indexType)),
+            equalTo(stringKey("elasticsearch.id"), experimental(id)),
+            equalTo(longKey("elasticsearch.version"), experimental(version))));
+  }
+
   protected boolean hasWriteVersion() {
     return true;
   }
 
-  protected void metricAssertion(String instrumentationName) throws Exception {
-    ClusterHealthStatus clusterHealthStatus =
-        testing.runWithSpan("parent", this::clusterHealthSync);
-
-    assertThat(clusterHealthStatus.name()).isEqualTo("GREEN");
-    testing.waitForTraces(1);
-
-    assertDurationMetric(testing, instrumentationName, DB_SYSTEM_NAME, DB_OPERATION_NAME);
+  protected boolean hasNetworkType() {
+    return false;
   }
 }
